Fix random number startup seed
Oh, the dangers of touching a PRG even slightly! I can hear Knuth tching!

The seed was set to the first byte of the serial number on start up. This failed because...
1. a seed to Marsaglia's XOR shifter [must never be zero](https://github.com/bigjosh/Move38-Arduino-Platform/blob/main/cores/blinklib/blinklib.cpp#L998-L1002)
2. the first byte of the serial number can (and *always* is on 328 chips!) be zero.

This patch instead uses the full serial number folded into a well-known base. This should never result in a zero seed and is more likely to produce a deterministic seed unique to each blink.